- Updated gmath memory allocation to use size_t type.
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@17738 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ lib/gmath/dalloc.c
@@ -12,7 +12,7 @@
  *  \return double * 
  */
 
- double *G_alloc_vector(int n)
+double *G_alloc_vector(size_t n)
 {
     return (double *) G_calloc (n, sizeof(double));
 }
@@ -52,7 +52,7 @@
  *  \return float * 
  */
 
- float *G_alloc_fvector(int n)
+float *G_alloc_fvector(size_t n)
 {
     return (float *) G_calloc (n, sizeof(float));
 }
